chore: organize datatable
