+++ .gitignore
@@ -50,3 +50,5 @@ playwright/.cache
 downloads
 .vite
 .nx
+.cursor/rules/nx-rules.mdc
+.github/instructions/nx.instructions.md
