+++ packages/ove/package.json
@@ -57,7 +57,6 @@
     "redux-thunk": "2.4.1",
     "remark-gfm": "^4.0.0",
     "reselect": "^4.1.7",
-    "shortid": "2.2.16",
     "tg-use-local-storage-state": "^16.0.3",
     "to-regex-range": "5.0.1",
     "use-debounce": "^8.0.4",
